Store actions as Configurations in Knative
This does some basic wiring up of `wsk action create` and `wsk action
list` to use Knative CRDs. You can create an action and subsequently
list it. Only the action's name and namespace are populated at the
moment.

We're not yet doing anything with the action's code nor creating any
kind of Knative build objects for the action.